+++ tests/plugins/lti_consumer/test_api.py
@@ -3,6 +3,7 @@ Tests for LTI Consumer plugin api
 """
 
 import json
+import random
 from unittest import mock
 
 from django.contrib.auth.models import AnonymousUser
@@ -44,13 +45,90 @@ class LtiConsumerPluginApiTestCase(TestCase):
         response = self.client.get(
             f"/api/v1.0/plugins/lti-consumer/{model_instance.pk}/context/"
         )
+        self.assertEqual(response.status_code, 200)
         content = json.loads(response.content)
 
-        self.assertEqual(response.status_code, 200)
         self.assertIn(content["url"], lti_consumer.url)
-        self.assertEqual(content["automatic_resizing"], lti_consumer.automatic_resizing)
+        self.assertTrue(content["is_automatic_resizing"])
         self.assertEqual(content["content_parameters"], "test_content")
         mock_params.assert_called_once()
+
+    @mock.patch.object(
+        LTIConsumer, "get_content_parameters", return_value="test_content"
+    )
+    def test_lti_consumer_api_get_context_is_automatic_resizing_default(
+        self, _mock_params
+    ):
+        """
+        The "is_automatic_resizing" parameter should default to True if it is not set
+        in the provider configuration.
+        """
+        placeholder = Placeholder.objects.create(slot="test")
+
+        lti_consumer = LTIConsumerFactory()
+        model_instance = add_plugin(
+            placeholder,
+            LTIConsumerPlugin,
+            "en",
+            url=lti_consumer.url,
+            lti_provider_id=lti_consumer.lti_provider_id,
+            is_automatic_resizing=lti_consumer.is_automatic_resizing,
+        )
+
+        # Set from settings
+        is_automatic_param = random.choice([True, False])
+        with override_settings(
+            RICHIE_LTI_PROVIDERS={
+                "lti_provider_test": {"is_automatic_resizing": is_automatic_param}
+            }
+        ):
+            response = self.client.get(
+                f"/api/v1.0/plugins/lti-consumer/{model_instance.pk}/context/"
+            )
+        self.assertEqual(response.status_code, 200)
+        content = json.loads(response.content)
+
+        self.assertEqual(content["is_automatic_resizing"], is_automatic_param)
+
+        # Set from the instance
+        with override_settings(RICHIE_LTI_PROVIDERS={"lti_provider_test": {}}):
+            response = self.client.get(
+                f"/api/v1.0/plugins/lti-consumer/{model_instance.pk}/context/"
+            )
+        self.assertEqual(response.status_code, 200)
+        content = json.loads(response.content)
+
+        self.assertTrue(content["is_automatic_resizing"])
+
+    @mock.patch.object(
+        LTIConsumer, "get_content_parameters", return_value="test_content"
+    )
+    def test_lti_consumer_api_get_context_is_automatic_resizing_manual(
+        self, _mock_params
+    ):
+        """
+        The "is_automatic_resizing" parameter should default to the value set in the
+        plugin model instance.
+        """
+        placeholder = Placeholder.objects.create(slot="test")
+
+        is_automatic = random.choice([True, False])
+        model_instance = add_plugin(
+            placeholder,
+            LTIConsumerPlugin,
+            "en",
+            lti_provider_id=None,
+            is_automatic_resizing=is_automatic,
+        )
+
+        with override_settings(RICHIE_LTI_PROVIDERS={"lti_provider_test": {}}):
+            response = self.client.get(
+                f"/api/v1.0/plugins/lti-consumer/{model_instance.pk}/context/"
+            )
+        self.assertEqual(response.status_code, 200)
+        content = json.loads(response.content)
+
+        self.assertEqual(content["is_automatic_resizing"], is_automatic)
 
     def test_lti_consumer_api_get_context_to_unknown_placeholder(self):
         """