+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add a new authentication backend `fonzie`
+- Add react-query to manage API requests and local data store
 
 ### Fixed