+++ CHANGELOG.md
@@ -15,6 +15,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 ### Added
 
 - Add a new authentication backend `fonzie`
+- Add react-query to manage API requests and local data store
 
 ### Changed