+++ CHANGELOG.md
@@ -19,6 +19,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
+- Allow configuring the inline ratio and auto-resizing for manually-defined
+  LTIConsumerPlugin instances
 - Use a unique name to LTIConsumerPlugin iframes to allow several plugins on the
 - Fix iframeResizer on LTIConsumer component when several plugins are on the
   same page