+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unrealeased]
 
+## [2.7.0] - 2021-06-04
+
 ### Added
 
 - Add "hreflang" meta to all pages to avoid duplicates in search engines
@@ -19,7 +21,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Fixed
 
-- Fix permissions on the LTI consumer plugin when granting institutor role
+- Fix permissions on the LTI consumer plugin when granting instructor role
 - Allow configuring the inline ratio and auto-resizing for manually-defined
   LTIConsumerPlugin instances
 - Use a unique name to LTIConsumerPlugin iframes to allow several plugins on the
@@ -1319,7 +1321,8 @@ us:
 - finish integrating the missing pages and improve the sandbox environment;
 - test and polish the use of richie as a django app / node dependency.
 
-[unreleased]: https://github.com/openfun/richie/compare/v2.6.0...master
+[unreleased]: https://github.com/openfun/richie/compare/v2.7.0...master
+[2.7.0]: https://github.com/openfun/richie/compare/v2.6.0...v2.7.0
 [2.6.0]: https://github.com/openfun/richie/compare/v2.5.0...v2.6.0
 [2.5.0]: https://github.com/openfun/richie/compare/v2.4.0...v2.5.0
 [2.4.0]: https://github.com/openfun/richie/compare/v2.3.3...v2.4.0